locationStandard util
